Assignment Completed
The page displays option to create username and password which is
dynamically updated in the mysql table and is shown in bootstrap html
table. There is option to delete the desired row of the table by
clicking on the username. Each row of the table can be updated by
clicking update button which leads to random username and password
generated and is reflected in the html table. I have kept the ID unique,
so even after deleting the a particular row, already assigned ID is not
given to other user.